fixed e2e cleanup

+++ tests/e2e/commands/cleanup.go
@@ -5,6 +5,7 @@ package commands
 
 import (
 	"encoding/json"
+	"fmt"
 	"os"
 	"os/user"
 	"path/filepath"
@@ -19,15 +20,20 @@ func DeleteE2EInventory() {
 	gomega.Expect(err).Should(gomega.BeNil())
 	homeDir := usr.HomeDir
 	inventoryE2E := filepath.Join(homeDir, constants.BaseDirName, "nodes/inventories/", constants.E2EClusterName)
+	fmt.Println("deleting: ", inventoryE2E)
 	err = os.RemoveAll(inventoryE2E)
 	gomega.Expect(err).Should(gomega.BeNil())
 }
 
 func DeleteNode(nodeID string) {
+	if nodeID == "" {
+		return
+	}
 	usr, err := user.Current()
 	gomega.Expect(err).Should(gomega.BeNil())
 	homeDir := usr.HomeDir
 	nodeE2E := filepath.Join(homeDir, constants.BaseDirName, "nodes", nodeID)
+	fmt.Println("deleting: ", nodeE2E)
 	err = os.RemoveAll(nodeE2E)
 	gomega.Expect(err).Should(gomega.BeNil())
 }
@@ -44,7 +50,7 @@ func DeleteE2ECluster() {
 	gomega.Expect(err).Should(gomega.BeNil())
 	gomega.Expect(clustersConfig.Clusters[constants.E2EClusterName]).To(gomega.Not(gomega.BeNil()))
 	clustersConfig.Clusters[constants.E2EClusterName] = models.ClusterConfig{}
-	content, err = json.Marshal(clustersConfig)
+	content, err = json.MarshalIndent(clustersConfig, "", "    ")
 	gomega.Expect(err).Should(gomega.BeNil())
 	err = os.WriteFile(filepath.Join(homeDir, constants.BaseDirName, relativePath, constants.ClustersConfigFileName), content, 0o600)
 	gomega.Expect(err).Should(gomega.BeNil())
